+++ android/src/main/java/com/odehbros/flutter_file_downloader/core/DownloadTaskBuilder.java
@@ -3,35 +3,45 @@ package com.odehbros.flutter_file_downloader.core;
 import android.app.Activity;
 
 public class DownloadTaskBuilder {
-    DownloadTaskBuilder instance;
-    String url, name;
+    String url, name, downloadDestination;
+    String notifications = "progressOnly";
     DownloadCallbacks callbacks;
     final Activity activity;
     private DownloadTask task;
 
     public DownloadTaskBuilder(final Activity activity) {
         this.activity = activity;
-        instance = this;
     }
 
     public DownloadTaskBuilder setUrl(final String url) {
-        instance.url = url;
-        return instance;
+        this.url = url;
+        return this;
     }
 
     public DownloadTaskBuilder setName(final String name) {
-        instance.name = name;
-        return instance;
+        this.name = name;
+        return this;
+    }
+
+    public DownloadTaskBuilder setShowNotifications(String notifications) {
+        if (notifications != null && !notifications.isEmpty())
+            this.notifications = notifications;
+        return this;
+    }
+
+    public DownloadTaskBuilder setDownloadDestination(final String downloadDestination) {
+        this.downloadDestination = downloadDestination;
+        return this;
     }
 
     public DownloadTaskBuilder setCallbacks(final DownloadCallbacks callbacks) {
-        instance.callbacks = callbacks;
-        return instance;
+        this.callbacks = callbacks;
+        return this;
     }
 
     public DownloadTask build() {
         if (task == null)
-            task = new DownloadTask(instance.activity, instance.url, instance.name, callbacks);
+            task = new DownloadTask(activity, url, name, notifications, downloadDestination, callbacks);
         try {
             return getDownloadTask();
         } catch (Exception e) {